refactor: Reorganize models and events directory structure
- Moved app_state and pane_state to models root for clearer hierarchy
- Moved yank_buffer to models/buffer where it logically belongs
- Renamed 'state' directory to 'coordinates' for geometry/position types
- Reorganized events: event bus/model/view events to models/events
- Moved event sources to io directory with other I/O abstractions
- Moved core types to models/pane_state/core for better encapsulation

This creates a cleaner architecture:
- models/ contains all data structures
- events are properly located in models/events
- I/O abstractions are consolidated in io/
- All tests passing (478 unit tests)
