+++ src/repl/models/app_state.rs
@@ -9,8 +9,7 @@
 //! - Views handle presentation and user interaction
 
 use crate::repl::events::{EventBus, ModelEvent, ViewEvent};
-use crate::repl::models::{MemoryYankBuffer, ResponseModel, StatusLine, YankBuffer};
-use super::pane_state::PaneState;
+use crate::repl::models::{MemoryYankBuffer, PaneState, ResponseModel, StatusLine, YankBuffer};
 use std::collections::HashMap;
 
 /// Type alias for event bus option to reduce complexity
@@ -23,26 +22,26 @@ type EventBusOption = Option<Box<dyn EventBus>>;
 pub struct AppState {
     // Core models
     pub response: ResponseModel,
-    
+
     // Pane states
     pub request_pane: PaneState,
     pub response_pane: PaneState,
     pub active_pane: crate::repl::events::Pane,
-    
+
     // Status line model
     pub status_line: StatusLine,
-    
+
     // HTTP session configuration
     pub http_session_headers: HashMap<String, String>,
-    
+
     // Event management
     pub event_bus: EventBusOption,
     pub pending_view_events: Vec<ViewEvent>,
     pub pending_model_events: Vec<ModelEvent>,
-    
+
     // Yank buffer for copy/paste operations
     pub yank_buffer: Box<dyn YankBuffer>,
-    
+
     // Configuration flags
     pub clipboard_enabled: bool,
     pub dcut_enabled: bool,
@@ -52,25 +51,25 @@ impl AppState {
     /// Create a new AppState with default values
     pub fn new() -> Self {
         use crate::repl::events::{Pane, PaneCapabilities};
-        
+
         // Default terminal size
         let (width, height) = (80, 24);
-        
+
         Self {
             response: ResponseModel::new(),
             request_pane: PaneState::new(
-                Pane::Request, 
-                width, 
-                height, 
+                Pane::Request,
+                width,
+                height,
                 false,
-                PaneCapabilities::FULL_ACCESS
+                PaneCapabilities::FULL_ACCESS,
             ),
             response_pane: PaneState::new(
-                Pane::Response, 
-                width, 
-                height, 
+                Pane::Response,
+                width,
+                height,
                 false,
-                PaneCapabilities::READ_ONLY | PaneCapabilities::SELECTABLE
+                PaneCapabilities::READ_ONLY | PaneCapabilities::SELECTABLE,
             ),
             active_pane: Pane::Request,
             status_line: StatusLine::new(),
@@ -83,13 +82,13 @@ impl AppState {
             dcut_enabled: true,
         }
     }
-    
+
     /// Set the event bus for this AppState
     pub fn set_event_bus(&mut self, event_bus: Box<dyn EventBus>) {
         self.event_bus = Some(event_bus);
         tracing::debug!("Event bus set for AppState");
     }
-    
+
     /// Get the active pane state
     pub fn active_pane_state(&self) -> &PaneState {
         match self.active_pane {
@@ -97,7 +96,7 @@ impl AppState {
             crate::repl::events::Pane::Response => &self.response_pane,
         }
     }
-    
+
     /// Get the active pane state mutably
     pub fn active_pane_state_mut(&mut self) -> &mut PaneState {
         match self.active_pane {
@@ -105,7 +104,7 @@ impl AppState {
             crate::repl::events::Pane::Response => &mut self.response_pane,
         }
     }
-    
+
     /// Get pane state by pane type
     pub fn get_pane_state(&self, pane: crate::repl::events::Pane) -> &PaneState {
         match pane {
@@ -113,7 +112,7 @@ impl AppState {
             crate::repl::events::Pane::Response => &self.response_pane,
         }
     }
-    
+
     /// Get pane state mutably by pane type
     pub fn get_pane_state_mut(&mut self, pane: crate::repl::events::Pane) -> &mut PaneState {
         match pane {
