+++ src/repl/models/pane_state/capabilities.rs
