+++ src/repl/models/pane_state/content.rs
@@ -6,7 +6,7 @@
 //! - Content manipulation with capability checking
 
 use crate::repl::events::{Pane, PaneCapabilities, ViewEvent};
-use crate::repl::models::state::geometry::Position;
+use crate::repl::models::coordinates::geometry::Position;
 use crate::repl::models::BufferModel;
 
 use super::PaneState;
