+++ src/repl/models/pane_state/core/mod.rs
@@ -1,0 +1,5 @@
+//! Core types for pane state
+
+mod types;
+
+pub use types::{EditorMode, LogicalPosition, LogicalRange, Pane, PaneCapabilities};
