+++ src/repl/models/pane_state/cursor_line.rs
@@ -7,7 +7,7 @@
 //! - Append mode positioning
 
 use crate::repl::events::{EditorMode, LogicalPosition, PaneCapabilities, ViewEvent};
-use crate::repl::models::state::geometry::Position;
+use crate::repl::models::coordinates::geometry::Position;
 
 use super::PaneState;
