+++ src/repl/models/pane_state/scrolling.rs
@@ -7,7 +7,7 @@
 //! - Mode-aware scrolling behavior
 
 use crate::repl::events::{EditorMode, LogicalPosition, PaneCapabilities, ViewEvent};
-use crate::repl::models::state::geometry::Position;
+use crate::repl::models::coordinates::geometry::Position;
 
 use super::{CursorMoveResult, PaneState, ScrollAdjustResult, ScrollResult};
