+++ src/repl/models/pane_state/visual_selection.rs
