+++ src/repl/models/pane_state/word_navigation.rs
@@ -7,7 +7,7 @@
 //! - Support for Japanese and multi-byte character word boundaries
 
 use crate::repl::events::{EditorMode, LogicalPosition, PaneCapabilities, ViewEvent};
-use crate::repl::models::state::geometry::Position;
+use crate::repl::models::coordinates::geometry::Position;
 
 use super::{OptionalPosition, PaneState};
