+++ src/repl/view_models/cursor_manager.rs
@@ -4,7 +4,7 @@
 //! This module provides high-level cursor operations that work with the current/other area abstraction.
 
 use crate::repl::events::LogicalPosition;
-use crate::repl::models::state::geometry::Position;
+use crate::repl::models::coordinates::geometry::Position;
 use crate::repl::view_models::core::ViewModel;
 use anyhow::Result;
