+++ src/repl/view_models/display_manager.rs
@@ -4,7 +4,7 @@
 //! This module coordinates between logical content and display representation.
 
 use crate::repl::events::{Pane, ViewEvent};
-use crate::repl::models::state::geometry::Position;
+use crate::repl::models::coordinates::geometry::Position;
 use crate::repl::models::DisplayCache;
 use crate::repl::view_models::core::{DisplayLineData, ViewModel};
